# app/openapi.yaml
openapi: 3.1.0
info:
  title: 論文調査メモ管理API
  version: 0.1.0
paths:
  /api/v1/papers:
    get:
      summary: 論文一覧を取得する
      parameters:
        - name: done
          in: query
          schema: { type: boolean }
          description: 完了状態で絞り込む
      responses:
        '200':
          description: タスク一覧
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Paper' }
    post:
      summary: 論文記録を新規作成する
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaperCreate' }
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Paper' }
        '422':
          description: バリデーションエラー
  /api/v1/papers/{paper_id}:
    get:
      summary: 論文メモを1件取得する
      parameters:
        - name: paper_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 論文メモ1件
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Paper' }
        '404':
          description: 論文メモが存在しない
    patch:
      summary: 論文メモを部分更新する
      parameters:
        - name: paper_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 更新1件
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Paper' }
        '204':
          description: 更新1件
        '404':
          description: 論文メモがない、もしくはパラメータ誤り
    delete:
      summary: 論文メモを削除する
      parameters:
        - name: paper_id
          in: path
          required: true
          schema: {type: integer }
      responses:
        '204':
          description: 削除1件
        '404':
          description: 該当論文メモがない。
    patch:
      summary: 論文メモを部分更新する
      parameters:
        - name: paper_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaperUpdate' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Paper' }
        '404':
          description: 論文メモが存在しない

components:
  schemas:
    PaperCreate:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 100 }
        published_date: { type: int, minimum: 1, maxmum: 9999, description: "西暦年。未来の年は指定できない", nullable: true }
        done: { type: boolean }
        memo: { type: string, nullable: true}
    Paper:
      type: object
      properties:
        id: { type: integer }
        title: { type: string, minLength: 1, maxLength: 100 }
        published_date: { type: int, minimum: 1, maxmum: 9999, description: "西暦年。未来の年は指定できない", nullable: true }
        done: { type: boolean }
        memo: { type: string, nullable: true}
        created_at: { type: string, format: date-time }
    PaperUpdate:
      type: object
      properties: 
        title: { type: string, minLength: 1, maxLength: 100 }
        published_date: { type: int, minimum: 1, maxmum: 9999, description: "西暦年。未来の年は指定できない", nullable: true }
        done: { type: boolean }
        memo: { type: string, nullable: true}
        created_at: { type: string, format: date-time }